Minimal code to demo a Flutter bug, which happens in Chrome on Android
when there is a TextButton inside Text.rich and the TextButton has
size smaller than the default minimal size (64×36) then the
TextButton is rendered incorrectly. Looks like it occupies an area
of default minimal size instead of the proper smaller real size area.

The compiled demo is published at
https://mshurkaeu-public.github.io/flutter-bugs-I-found-or-reported/04-TextButton-with-a-given-size-inside-TextSpan-in-Android-browser/

The URL is displayed as expected in desktop Chrome 112.0.5615.165 (Official Build) (64-bit)
And is displayed incorrectly in Android Chrome 112.0.5615.136

To build the demo I used command
flutter build --verbose web --release --web-renderer canvaskit --base-href "/flutter-bugs-I-found-or-reported/04-TextButton-with-a-given-size-inside-TextSpan-in-Android-browser/"

Reported to the Flutter team as https://github.com/flutter/flutter/issues/125756