Minimal code to demo a Flutter bug, which happens in Chrome on Android
when there is a full-width WidgetSpan inside Text.rich and text of the
TextSpan, following the WidgetSpan, starts with newline characters then
the newlines characters are "eaten". Not renedered. The bug appears only
with canvaskit web renderer. The html renderer produces result, which
behaves as expected.

The compiled demo of the bug is published at
canvaskit - https://mshurkaeu-public.github.io/flutter-bugs-I-found-or-reported/05-canvaskit-newlines-are-eaten-after-WidgetSpan-in-Android-browser/
html - https://mshurkaeu-public.github.io/flutter-bugs-I-found-or-reported/05-html-newlines-are-eaten-after-WidgetSpan-in-Android-browser/

The URLs are displayed as expected in desktop Chrome 113.0.5672.126 (Official Build) (64-bit)
And are displayed incorrectly in Android Chrome 113.0.5672.162

It seems to be a manifistation of the bug, already reported to the
Flutter team - https://github.com/flutter/flutter/issues/126258